+++ src/managed/mod.rs
@@ -288,9 +288,10 @@ impl<T, E> Pool<T, E> {
     }
     /// Retrieve status of the pool
     pub fn status(&self) -> Status {
+        let max_size = self.inner.config.max_size;
         let size = self.inner.size.load(Ordering::Relaxed);
         let available = self.inner.available.load(Ordering::Relaxed);
-        Status { size, available }
+        Status { max_size, size, available }
     }
 }